timeularex v0.1.2 Timeularex

NOTE: Timeularex is in early stages of development. Use in production should be considered with caution.

Timeularex is an API client for the Timeular public API

Timular is a service to improve time-tracking of activities.

Configuration

The client provides two options for configuration. The first involves the typical setting of variables in your config.exs file:

config :timeularex,
  api_key: <<API_KEY>>,
  api_secret: <<API_SECRET>>,
  api_url: "https://api.timeular.com/api/v2"

Additionally, you can utilize api_key/1 and api_secret/1 functions in the TImeularex.Config module.

Your API key and secret can be retrieved from your account app settings

Functions

activate_device(device_serial)

Sets the status of a device to active.

activities()

List all activities associated with the connected account.

archive_activity(activity_id)

Archive an Activity. Time tracked with the Activity will be preserved.

archived_activities()

List all archived Activities.

assign_side_activity(device_side, activity_id)

Assign an Activity to a given device side.

create_activity(activity)

Create a new Activity. The activity should have a name and color. A Name doesn’t have to be unique. Optionally, you can also provide an Integration to which the activity will belong to. You can obtain list of enabled Integrations with integrations/0.

deactivate_device(device_serial)

Sets the status of a device to inactive.

devices()

List all devices.

disable_device(device_serial)

Disable a device.

edit_tracking_note(activity_id, note)

Edit the notes associated an Activity being tracked.

enable_device(device_serial)

Enable a device.

init(_)

integrations()

List all the integrations associated with the connected account.

remove_device(device_serial)

Remove a device from the list of known devices. Use activate_device/1 to make the device active again.

reports(start_timestamp, stop_timestamp)

start(type, args)

Callback implementation for Application.start/2.

start_tracking(activity_id)

Start tracking a given Activity.

stop_tracking(activity_id)

Stop tracking a given Activity.

tags_and_mentions()

time_entries(stopped_after, started_before)

Return all time entries that falls between a given time range.

time_entry(entry_id)

Return a time entry by ID.

tracking()

Returns information regarding what is currently being tracked.

unassign_side_activity(device_side, activity_id)

Unassigns an Activity associated with a given device side.

update_activity(activity)

Update an existing Activity's name and/or color.

update_device_name(device_serial, name)

Update the name of a device.